services/doc-pipeline/requirements.txt
azure-functions>=1.21.0,<2.0.0
azure-storage-blob>=12.24.0,<13.0.0
azure-ai-formrecognizer>=3.3.0,<4.0.0
azure-search-documents>=11.6.0,<12.0.0
azure-cosmos>=4.9.0,<5.0.0
azure-storage-queue>=12.12.0,<13.0.0
azure-identity>=1.19.0,<2.0.0
opentelemetry-api>=1.29.0,<2.0.0
pydantic>=2.10.0,<3.0.0
httpx>=0.28.0,<1.0.0
